# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AImpactScanner API
  version: 1.0.0
  description: |
    Analyze websites for AI search visibility using the MASTERY-AI Framework v3.1.1.

    ## Authentication

    All endpoints require authentication via one of:
    - **Bearer token**: JWT obtained from Supabase Auth (`Authorization: Bearer <token>`)
    - **API key**: Passed in the `X-API-Key` header

    ## Rate Limits

    Rate limits are per-day and vary by subscription tier:

    | Tier   | Calls/day |
    |--------|-----------|
    | Free   | 10        |
    | Coffee | 100       |
    | Growth | 1,000     |
    | Scale  | 10,000    |

    Response headers on every request:
    - `X-RateLimit-Limit` -- daily quota for your tier
    - `X-RateLimit-Remaining` -- approximate remaining calls
    - `X-RateLimit-Reset` -- Unix timestamp when the limit resets

    ## Monthly Analysis Limits

    In addition to API rate limits, scan submissions (`POST /v1/scan`) are capped
    per calendar month:

    | Tier   | Scans/month |
    |--------|-------------|
    | Free   | 3           |
    | Coffee | 10          |
    | Growth | 40          |
    | Scale  | 100         |
  contact:
    name: AI Search Mastery
    url: https://aimpactscanner.com
servers:
  - url: https://aimpactscanner-backend-production.up.railway.app
    description: Production
  - url: http://localhost:3001
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/webhooks:
    post:
      operationId: createWebhook
      tags: [Webhooks]
      summary: Create a new webhook
      description: |
        Register a URL to receive event notifications. The response includes
        the webhook secret which is used to verify delivery signatures.
        Save it immediately -- it will be masked in subsequent responses.
        Maximum 10 webhooks per user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS endpoint to receive events (HTTP allowed for localhost in development)
                  example: https://api.example.com/webhooks/aimpact
                events:
                  type: array
                  items:
                    $ref: "#/components/schemas/WebhookEvent"
                  minItems: 1
                  example: [scan.complete]
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook:
                    $ref: "#/components/schemas/WebhookConfig"
                  warning:
                    type: string
                    example: Save the secret now. It will be masked in subsequent responses.
        "400":
          description: Validation error or limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listWebhooks
      tags: [Webhooks]
      summary: List your webhooks
      description: Returns all webhooks for the authenticated user. Secrets are masked.
      responses:
        "200":
          description: Webhook list
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/webhooks/{id}:
    parameters:
      - $ref: "#/components/parameters/WebhookId"

    get:
      operationId: getWebhook
      tags: [Webhooks]
      summary: Get a webhook with recent deliveries
      description: Returns the webhook configuration (secret masked) and its 10 most recent delivery attempts.
      responses:
        "200":
          description: Webhook details
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook:
                    $ref: "#/components/schemas/WebhookConfig"
                  recent_deliveries:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookDelivery"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateWebhook
      tags: [Webhooks]
      summary: Update a webhook
      description: Update URL, subscribed events, or active status. Provide only the fields you want to change.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    $ref: "#/components/schemas/WebhookEvent"
                  minItems: 1
                is_active:
                  type: boolean
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook:
                    $ref: "#/components/schemas/WebhookConfig"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteWebhook
      tags: [Webhooks]
      summary: Delete a webhook
      responses:
        "200":
          description: Webhook deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Webhook deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/webhooks/{id}/test:
    parameters:
      - $ref: "#/components/parameters/WebhookId"
    post:
      operationId: testWebhook
      tags: [Webhooks]
      summary: Send a test event to a webhook
      description: Delivers a synthetic `scan.complete` event to the webhook endpoint so you can verify your integration.
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  delivery_id:
                    type: string
                    format: uuid
                  status_code:
                    type: integer
                  error:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key issued from your dashboard
  parameters:
...
    WebhookId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Webhook identifier
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: Unauthorized
            code: UNAUTHORIZED
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: Not found
            code: NOT_FOUND
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error: An unexpected error occurred
            code: INTERNAL_ERROR
  schemas:
...
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
          description: Machine-readable error code
...
    WebhookEvent:
      type: string
      enum:
        - scan.complete
        - llmstxt.drift_detected
        - monitor.citation_found
        - monitor.citation_lost
    WebhookConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: "#/components/schemas/WebhookEvent"
        secret:
          type: string
          description: HMAC-SHA256 signing secret (masked after creation)
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookDelivery:
      type: object
      properties:
        id:
          type: string
          format: uuid
        webhook_id:
          type: string
          format: uuid
        event:
          $ref: "#/components/schemas/WebhookEvent"
        payload:
          type: object
        status:
          type: string
          enum: [pending, success, failed]
        status_code:
          type: integer
          nullable: true
        response_body:
          type: string
          nullable: true
        attempts:
          type: integer
        max_attempts:
          type: integer
        next_retry_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true